backend/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.2
django-allauth==0.49.0
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-filter==23.4
django-js-asset==2.0.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
joblib==1.3.2
MarkupSafe==2.1.0
nltk==3.8.1
oauthlib==3.2.0
openapi-codec==1.3.2
packaging==21.3
Pillow==9.0.1
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2021.3
regex==2023.10.3
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tqdm==4.66.1
Unipath==1.1
uritemplate==4.1.1
urllib3==1.26.8
